requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
python-binance